requirements.txt
googletrans==4.0.0-rc1
vosk
webvtt-py
flask
uvicorn
gunicorn
